# Speed up `replaceRegexp(All|One)` if the pattern is trivial