--- a/tests/queries/0_stateless/02864_replace_regexp_string_fallback.reference
+++ b/tests/queries/0_stateless/02864_replace_regexp_string_fallback.reference
@@ -1,0 +1,1 @@
+Hello	l	x	Hexlo	Hexxo