The idea is to transform `replaceRegexp(One|All)` into `replace(One|All)` if the pattern is trivial.

This PR is a continuation of #62436 but compared to the original implementation, `ReplaceRegexpImpl` keeps using re2 instead of OptimizedRegularExpression to keep the code overall simple and maintainable.

Measurements with haystack `Many years later as he faced the firing squad, Colonel Aureliano Buendia was to remember that distant afternoon when his father took him to discover ice.` on 5 mio rows

- and complex pattern `\s+` and replacement `\\0\n`
    - replaceRegexpAll: 13.67 sec
    - replaceRegexpOne: 0.628 sec

- and trivial pattern ` ` (space) and replacement `\n`:
    - replaceRegexpAll: 1.63 sec
    - replaceRegexpOne: 0.16 sec

Courtesy to @taiyang-li !

### Changelog category (leave one):
- Performance Improvement

### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Functions `replaceRegexpAll` and `replaceRegexpOne` are now significantly faster if the pattern is trivial, i.e. contains no metacharacters, pattern classes, flags, grouping characters etc. (Thanks to Taiyang Li)