@divanik 2024-07-08
static void vectorFixedConstantConstant(...) - why can't we apply the same optimisation here?