@rschu1ze 2024-07-08
We could, in theory. The shortcut is deliberately implemented only for the most frequent case of non-const patterns & const needle & const replacement. The other functions (including `vectorFixedConstantConstant()`) handle special cases that are not really relevant in practice, e.g. non-const pattern & non-const needle & non-const replacement. Afaik FixedString is also sort of exotic.

The code contains a comment:
```cpp
/// Try to use non-regexp string replacement. This shortcut is implemented only for const-needles + const-replacement as
        /// pattern analysis incurs some cost too.
`