@rschu1ze 2024-07-08
Perf test failure is because test `replaceRegexpFallback` only been introduced with this PR.